This is a rework of the report portal Spock agent using the updated client pieces.
