+++ build.gradle
@@ -17,13 +17,14 @@ repositories {
 }
 
 dependencies {
-  compile 'com.epam.reportportal:client-java-core:2.7.1'
-  
-  compileOnly 'org.spockframework:spock-core:1.0-groovy-2.0'
+  compile 'com.epam.reportportal:client-java:4.0.10'
+  compile 'com.epam.reportportal:commons-model:4.2.0'
+
+  compileOnly ('org.spockframework:spock-core:1.1-groovy-2.4')
 
   testCompile 'org.hamcrest:hamcrest-all:1.3'
   testCompile 'junit:junit:4.12'
-  testCompile 'org.spockframework:spock-core:1.0-groovy-2.0'
+  testCompile ('org.spockframework:spock-core:1.1-groovy-2.4')
   testCompile 'org.mockito:mockito-all:2.0.2-beta'
   testCompile 'pl.pragmatists:JUnitParams:1.0.4'
   testCompile 'org.apache.commons:commons-lang3:3.5'
