+++ src/main/java/com/epam/reportportal/spock/FixtureFootprint.java
@@ -15,6 +15,7 @@
  */
 package com.epam.reportportal.spock;
 
+import io.reactivex.Maybe;
 import org.spockframework.runtime.model.MethodInfo;
 
 /**
@@ -24,12 +25,12 @@ import org.spockframework.runtime.model.MethodInfo;
  */
 class FixtureFootprint extends ReportableItemFootprint<MethodInfo> {
 
-	FixtureFootprint(MethodInfo item, String id) {
-		super(item, id);
-	}
+    FixtureFootprint(MethodInfo item, Maybe<String> id) {
+        super(item, id);
+    }
 
-	@Override
-	public boolean hasDescendants() {
-		return false;
-	}
+    @Override
+    public boolean hasDescendants() {
+        return false;
+    }
 }
