+++ src/main/java/com/epam/reportportal/spock/NodeFootprint.java
@@ -15,19 +15,18 @@
  */
 package com.epam.reportportal.spock;
 
-import static com.google.common.base.Predicates.and;
-import static com.google.common.collect.Iterables.find;
-import static com.google.common.collect.Lists.newArrayList;
-import static com.google.common.collect.Lists.newArrayListWithCapacity;
+import static rp.com.google.common.base.Predicates.and;
+import static rp.com.google.common.collect.Iterables.find;
+import static rp.com.google.common.collect.Lists.newArrayList;
+import static rp.com.google.common.collect.Lists.newArrayListWithCapacity;
 
 import java.util.List;
-
-import javax.annotation.Nullable;
-
+import io.reactivex.Maybe;
+import org.spockframework.util.Nullable;
 import org.spockframework.runtime.model.MethodInfo;
 import org.spockframework.runtime.model.NodeInfo;
+import rp.com.google.common.base.Predicate;
 
-import com.google.common.base.Predicate;
 
 /**
  * Abstract entity for the representation of the metadata for the reportable
@@ -37,52 +36,51 @@ import com.google.common.base.Predicate;
  */
 abstract class NodeFootprint<T extends NodeInfo> extends ReportableItemFootprint<T> {
 
-	/*
-	 * Approximate fixtures count, which should match the most of cases. This is
-	 * a some kind of "happy medium" between memory consumption and potential
-	 * performance drawback on arrays coping
-	 */
-	private static final int APPROXIMATE_CAPACITY = 4;
+    /*
+     * Approximate fixtures count, which should match most cases. This is
+     * a kind of "happy medium" between memory consumption and potential
+     * performance drawback on arrays coping
+     */
+    private static final int APPROXIMATE_CAPACITY = 4;
 
-	private final List<ReportableItemFootprint<MethodInfo>> fixtures;
+    private final List<ReportableItemFootprint<MethodInfo>> fixtures;
 
-	NodeFootprint(T nodeInfo, String id) {
-		super(nodeInfo, id);
-		fixtures = newArrayListWithCapacity(APPROXIMATE_CAPACITY);
-	}
+    NodeFootprint(T nodeInfo, Maybe<String> id) {
+        super(nodeInfo, id);
+        fixtures = newArrayListWithCapacity(APPROXIMATE_CAPACITY);
+    }
 
+    ReportableItemFootprint<MethodInfo> findFixtureFootprint(final MethodInfo fixture) {
+        Predicate<ReportableItemFootprint<MethodInfo>> criteria = createFixtureMatchPredicate(fixture);
+        return find(getFixtures(), criteria);
+    }
 
-	ReportableItemFootprint<MethodInfo> findFixtureFootprint(final MethodInfo fixture) {
-		Predicate<ReportableItemFootprint<MethodInfo>> criteria = createFixtureMatchPredicate(fixture);
-		return find(getFixtures(), criteria);
-	}
+    /**
+     * Find unpublished fixture footprint. It used to address an issue, when the node footprint can have multiple fixture
+     * footprint, wrapping the same {@link MethodInfo}
+     *
+     * @param fixture target method info
+     * @return footprint
+     */
+    ReportableItemFootprint<MethodInfo> findUnpublishedFixtureFootprint(final MethodInfo fixture) {
+        Predicate<ReportableItemFootprint<MethodInfo>> criteria = and(createFixtureMatchPredicate(fixture), IS_NOT_PUBLISHED);
+        return find(getFixtures(), criteria);
+    }
 
-	/**
-	 * Find unpublished fixture footprint. It used to address an issue, when the node footprint can have multiple fixture
-	 * footprint, wrapping the same {@link MethodInfo}
-	 * 
-	 * @param fixture target method info
-	 * @return footprint
-	 */
-	ReportableItemFootprint<MethodInfo> findUnpublishedFixtureFootprint(final MethodInfo fixture) {
-		Predicate<ReportableItemFootprint<MethodInfo>> criteria = and(createFixtureMatchPredicate(fixture), IS_NOT_PUBLISHED);
-		return find(getFixtures(), criteria);
-	}
+    void addFixtureFootprint(FixtureFootprint footprint) {
+        fixtures.add(footprint);
+    }
 
-	void addFixtureFootprint(FixtureFootprint footprint) {
-		fixtures.add(footprint);
-	}
+    List<ReportableItemFootprint<MethodInfo>> getFixtures() {
+        return newArrayList(fixtures);
+    }
 
-	private List<ReportableItemFootprint<MethodInfo>> getFixtures() {
-		return newArrayList(fixtures);
-	}
-
-	private static Predicate<ReportableItemFootprint<MethodInfo>> createFixtureMatchPredicate(final MethodInfo fixture) {
-		return new Predicate<ReportableItemFootprint<MethodInfo>>() {
-			@Override
-			public boolean apply(@Nullable ReportableItemFootprint<MethodInfo> footprint) {
-				return footprint != null && fixture.equals(footprint.getItem());
-			}
-		};
-	}
+    private static Predicate<ReportableItemFootprint<MethodInfo>> createFixtureMatchPredicate(final MethodInfo fixture) {
+        return new Predicate<ReportableItemFootprint<MethodInfo>>() {
+            @Override
+            public boolean apply(@Nullable ReportableItemFootprint<MethodInfo> footprint) {
+                return footprint != null && fixture.equals(footprint.getItem());
+            }
+        };
+    }
 }
