--- src/main/java/com/epam/reportportal/spock/SpockListenersModule.java
@@ -1,49 +1,0 @@
-/*
- * Copyright (C) 2019 EPAM Systems
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.epam.reportportal.spock;
-
-import javax.inject.Singleton;
-
-import org.spockframework.runtime.IRunListener;
-import org.spockframework.runtime.extension.IMethodInterceptor;
-
-import com.google.inject.AbstractModule;
-import com.google.inject.Provides;
-
-/**
- * <a href="https://github.com/google/guice">Guice</a> module definition for the Spock agent
- *
- * @author Dzmitry Mikhievich
- */
-class SpockListenersModule extends AbstractModule {
-
-	@Override
-	protected void configure() {
-		bind(ISpockReporter.class).toProvider(SpockReporterProvider.class).asEagerSingleton();
-		bind(IRunListener.class).to(ReportableRunListener.class).asEagerSingleton();
-	}
-
-	@Provides
-	private AbstractLaunchContext launchContext() {
-		return new LaunchContextImpl();
-	}
-
-	@Provides
-	@Singleton
-	IMethodInterceptor fixturesInterceptor(ISpockReporter spockReporter) {
-		return new FixtureInterceptor(spockReporter);
-	}
-}
